Make formal argument names more consistent.
srfi-231.scm:

1.  Make formal variablle names more consistent among procedure declarations.  The following exceptions are intentional and meant to prod the reader to look at the arguments more closely:

array-decurry[!] and array-block[!]:  The arguments are specifically arrays of arrays, so we use a different notation for the first argument ("A" instead of "array").

array-inner-product:  too much to explain here, making notation "consistent" hurts, not helps.

make-interval: The handling of arguments is different, so usr different notation.

srfi-231.html:

1.  Regenerate from srfi-231.scm.